+++ examples/mplapack/00_GeneralLinearEquations/Rgesvx_example_mpfr.cpp
@@ -96,7 +96,7 @@ int main() {
     equed[0] = 'N';
     equed[1] = '\0';
     mpfr_class rcond;
-    for (mplapackint j = 0; j < n; j++) for (mplapackint i = 0; i < n; i++) a[i + j * lda] = mpfr_class(1.0) / mpfr_class(i + j + 1);
+    for (mplapackint j = 0; j < n; j++) for (mplapackint i = 0; i < n; i++) a[i + j * lda] = mpfr_class(1.0) / mpfr_class((double)(i + j + 1));
     for (mplapackint i = 0; i < n; i++) xexact[i] = (i % 2 == 0) ? mpfr_class(1.0) : mpfr_class(-1.0);
     for (mplapackint i = 0; i < n; i++) {
         b[i] = 0.0;
