+++ examples/mplapack/00_GeneralLinearEquations/generic/Rgesvx_example_generic.cpp
@@ -48,7 +48,7 @@ int main() {
     equed[0] = 'N';
     equed[1] = '\0';
     REAL rcond;
-    for (INTEGER j = 0; j < n; j++) for (INTEGER i = 0; i < n; i++) a[i + j * lda] = REAL(1.0) / REAL(i + j + 1);
+    for (INTEGER j = 0; j < n; j++) for (INTEGER i = 0; i < n; i++) a[i + j * lda] = REAL(1.0) / REAL((double)(i + j + 1));
     for (INTEGER i = 0; i < n; i++) xexact[i] = (i % 2 == 0) ? REAL(1.0) : REAL(-1.0);
     for (INTEGER i = 0; i < n; i++) {
         b[i] = 0.0;
