+++ examples/mplapack/04_NonsymmetricEigenproblems/Cgeev_NPR_binary128.cpp
@@ -144,7 +144,7 @@ int main() {
     std::complex<mplapack_binary128_t> _pi = pi(mplapack_binary128_t(0.0));
     std::complex<mplapack_binary128_t> *lambda = new std::complex<mplapack_binary128_t>[n];
     for (int h = 1; h <= n; h++) {
-        lambda [h - 1] = delta + std::complex<mplapack_binary128_t>(2.0, 0.0) * sqrt (sigma * tau) * cos( (mplapack_binary128_t(h) * _pi) / mplapack_binary128_t((int)n + 1) );
+        lambda [h - 1] = delta + std::complex<mplapack_binary128_t>(2.0, 0.0) * sqrt (sigma * tau) * cos( (mplapack_binary128_t((double)h) * _pi) / mplapack_binary128_t((double)((int)n + 1)) );
     }
     printf("lambda_true = "); printvec(lambda, n); printf("\n");
     printf("vr ="); printmat(n,n,vr,n); printf("\n");
