+++ examples/mplapack/90_PrecisionComparison/Gejsv_vs_gesvd_binary128.cpp
@@ -68,8 +68,8 @@ int main() {
     mplapack_binary128_t *b = new mplapack_binary128_t[m * n];
     for (mplapackint j = 0; j < n; j++)
         for (mplapackint i = 0; i < m; i++) {
-            mplapack_binary128_t scale = pow(mplapack_binary128_t(10.0), mplapack_binary128_t(i - 4));
-            a[i + j * lda] = scale / (mplapack_binary128_t(i + j + 1));
+            mplapack_binary128_t scale = pow(mplapack_binary128_t(10.0), mplapack_binary128_t((double)(i - 4)));
+            a[i + j * lda] = scale / (mplapack_binary128_t((double)(i + j + 1)));
             b[i + j * lda] = a[i + j * lda];
         }
     mplapack_binary128_t *s1 = new mplapack_binary128_t[n];
